+++ src/SoundDeck.Core/Sessions/ProcessNamePredicate.cs
@@ -1,6 +1,7 @@
 ﻿namespace SoundDeck.Core.Sessions
 {
     using System;
+    using System.Collections.Concurrent;
     using System.Diagnostics;
     using NAudio.CoreAudioApi;
     using SoundDeck.Core.Extensions;
@@ -13,6 +14,11 @@
     public class ProcessNamePredicate : ISessionPredicate
     {
         /// <summary>
+        /// Provides a cache of invalid <see cref="AppInfo.GetFromAppUserModelId(string)"/> elements.
+        /// </summary>
+        private static readonly ConcurrentDictionary<string, bool> INVALID_APP_USER_MODEL_IDS = new ConcurrentDictionary<string, bool>();
+
+        /// <summary>
         /// Initializes a new instance of the <see cref="ProcessNamePredicate"/> class.
         /// </summary>
         /// <param name="processName">Name of the process to match against.</param>
@@ -23,6 +29,22 @@
         /// Gets the name of the process to match against.
         /// </summary>
         public string ProcessName { get; }
+
+        /// <inheritdoc/>
+        public bool Equals(ISessionPredicate x, ISessionPredicate y)
+        {
+            if (x is ProcessNamePredicate a
+                && y is ProcessNamePredicate b)
+            {
+                return a.ProcessName.Equals(b.ProcessName, StringComparison.OrdinalIgnoreCase);
+            }
+
+            return x == null && y == null;
+        }
+
+        /// <inheritdoc/>
+        public int GetHashCode(ISessionPredicate obj)
+            => this.ProcessName.GetHashCode();
 
         /// <inheritdoc/>
         public bool IsMatch(AudioSessionControl session)
@@ -47,10 +69,11 @@
             try
             {
                 return session.SourceAppUserModelId.Contains(this.ProcessName, StringComparison.OrdinalIgnoreCase)
-                    || AppInfo.GetFromAppUserModelId(session.SourceAppUserModelId).DisplayInfo.DisplayName.Contains(this.ProcessName, StringComparison.OrdinalIgnoreCase);
+                    || (!INVALID_APP_USER_MODEL_IDS.TryGetValue(session.SourceAppUserModelId, out var _) && AppInfo.GetFromAppUserModelId(session.SourceAppUserModelId).DisplayInfo.DisplayName.Contains(this.ProcessName, StringComparison.OrdinalIgnoreCase));
             }
             catch
             {
+                INVALID_APP_USER_MODEL_IDS.TryAdd(session.SourceAppUserModelId, false);
                 return false;
             }
         }