+++ src/SoundDeck.Plugin/Models/Settings/AppAudioControlsSettings.cs
@@ -1,0 +1,6 @@
+﻿namespace SoundDeck.Plugin.Models.Settings
+{
+    public class AppAudioControlSettings : ProcessSelectionCriteriaSettings
+    {
+    }
+}